# services/learning/api/http/v1/api.yaml
openapi: 3.0.3
info:
  title: even-learning
  version: 0.2.0
  description: Student learning — courses, lessons, progress, review, dictionary
paths:
...
  /api/v1/courses/{courseId}:
    get:
      operationId: getCourse
      tags: [Student, Courses]
      summary: Get course details (requires enrollment)
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/DefaultError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    DefaultError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        error:
          type: string
...
    Course:
      type: object
      required: [id, title, target_language_id, ui_language_id, owner_id, is_published]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        target_language_id:
          type: string
          format: uuid
        ui_language_id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        is_published:
          type: boolean
...